Increase precision of .ts timestamps.

+++ tsreduce.c
@@ -158,7 +158,7 @@ int main( int argc, char *argv[] )
     else if (argc == 3 && strcmp(argv[1], "repeats") == 0)
         return detect_repeats(argv[2]);
     else if (argc == 7 && strcmp(argv[1], "bjd") == 0)
-        return calculate_bjd(argv[2], argv[3], argv[4], argv[5], atoi(argv[6]));
+        return calculate_bjd(argv[2], argv[3], argv[4], argv[5], atof(argv[6]));
     else
         error("Invalid args");
     return 0;
